+++ gwcelery/tests/test_tools_nagios.py
@@ -22,7 +22,8 @@ def test_nagios_unknown_error(monkeypatch, capsys):
 
 def test_nagios(capsys, monkeypatch, socket_enabled, starter):
     mock_lvalert_client = Mock()
-    monkeypatch.setattr('sleek_lvalert.LVAlertClient', mock_lvalert_client)
+    monkeypatch.setattr(
+        'gwcelery.lvalert.client.LVAlertClient', mock_lvalert_client)
     unix_socket = app.conf['broker_url'].replace('redis+socket://', '')
 
     # no broker
@@ -49,23 +50,13 @@ def test_nagios(capsys, monkeypatch, socket_enabled, starter):
     out, err = capsys.readouterr()
     assert 'CRITICAL: Not all expected queues are active' in out
 
-    # worker, no tasks
+    # worker, no LVAlert nodes
 
     starter.python_process(
         args=(['gwcelery', 'worker', '-l', 'info', '--pool', 'solo',
                '-Q', 'celery,exttrig,openmp,superevent,voevent'],),
         target=app.start, timeout=10, magic_words=b'ready.')
 
-    with pytest.raises(SystemExit) as excinfo:
-        app.start(['gwcelery', 'nagios'])
-    assert excinfo.value.code == nagios.NagiosPluginStatus.CRITICAL
-    out, err = capsys.readouterr()
-    assert 'CRITICAL: Not all expected tasks are active' in out
-
-    # tasks, no LVAlert nodes
-
-    monkeypatch.setattr('gwcelery.tools.nagios.get_active_tasks',
-                        lambda _: nagios.get_expected_tasks(app))
     mock_lvalert_client.configure_mock(**{
         'return_value.get_subscriptions.return_value': {}})
 
@@ -77,9 +68,10 @@ def test_nagios(capsys, monkeypatch, socket_enabled, starter):
 
     # tasks, too many LVAlert nodes
 
-    mock_lvalert_client.configure_mock(**{
-        'return_value.get_subscriptions.return_value':
-        nagios.get_expected_lvalert_nodes() | {'foobar'}})
+    expected_lvalert_nodes = nagios.get_expected_lvalert_nodes(app)
+    monkeypatch.setattr(
+        'celery.app.control.Inspect.stats', Mock(return_value={'foo': {
+            'lvalert-nodes': expected_lvalert_nodes | {'foobar'}}}))
 
     with pytest.raises(SystemExit) as excinfo:
         app.start(['gwcelery', 'nagios'])
@@ -89,9 +81,9 @@ def test_nagios(capsys, monkeypatch, socket_enabled, starter):
 
     # LVAlert nodes present, no VOEvent broker peers
 
-    mock_lvalert_client.configure_mock(**{
-        'return_value.get_subscriptions.return_value':
-        nagios.get_expected_lvalert_nodes()})
+    monkeypatch.setattr(
+        'celery.app.control.Inspect.stats', Mock(return_value={'foo': {
+            'lvalert-nodes': expected_lvalert_nodes}}))
 
     with pytest.raises(SystemExit) as excinfo:
         app.start(['gwcelery', 'nagios'])
@@ -102,8 +94,9 @@ def test_nagios(capsys, monkeypatch, socket_enabled, starter):
     # VOEvent broker peers, no VOEvent receiver peers
 
     monkeypatch.setattr(
-        'celery.app.control.Inspect.stats',
-        Mock(return_value={'foo': {'voevent-broker-peers': ['127.0.0.1']}}))
+        'celery.app.control.Inspect.stats', Mock(return_value={'foo': {
+            'voevent-broker-peers': ['127.0.0.1'],
+            'lvalert-nodes': expected_lvalert_nodes}}))
 
     with pytest.raises(SystemExit) as excinfo:
         app.start(['gwcelery', 'nagios'])
@@ -116,7 +109,8 @@ def test_nagios(capsys, monkeypatch, socket_enabled, starter):
     monkeypatch.setattr(
         'celery.app.control.Inspect.stats',
         Mock(return_value={'foo': {'voevent-broker-peers': ['127.0.0.1'],
-                                   'voevent-receiver-peers': ['127.0.0.1']}}))
+                                   'voevent-receiver-peers': ['127.0.0.1'],
+                                   'lvalert-nodes': expected_lvalert_nodes}}))
 
     with pytest.raises(SystemExit) as excinfo:
         app.start(['gwcelery', 'nagios'])